refactor(message): migrate message-related classes to a modular structure and improve code organization
