--- Heart/Message/Domain/Repositories/MessageRepository.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Heart\Message\Domain\Repositories;
-
-use Heart\Message\Domain\DTO\NewMessageDTO;
-use Heart\Message\Domain\Entities\MessageEntity;
-
-interface MessageRepository
-{
-    public function create(NewMessageDTO $payload, string $providerId, int $obtainedExperience): MessageEntity;
-}
