--- Heart/Message/MessageDomain.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Heart\Message;
-
-use Heart\Core\Contracts\DomainInterface;
-use Heart\Message\Infrastructure\Providers\MessageRouteProvider;
-use Heart\Message\Infrastructure\Providers\MessageServiceProvider;
-
-class MessageDomain extends DomainInterface
-{
-    public function registerProvider(): array
-    {
-        return [
-            MessageServiceProvider::class,
-            MessageRouteProvider::class,
-        ];
-    }
-}
