+++ app-modules/message/composer.json
@@ -1,0 +1,27 @@
+{
+	"name": "he4rt/message",
+	"description": "Message module for He4rt community bot",
+	"type": "library",
+	"version": "1.0",
+	"license": "proprietary",
+	"require": {},
+	"autoload": {
+		"psr-4": {
+			"He4rt\\Message\\": "src/",
+			"He4rt\\Message\\Database\\Factories\\": "database/factories/"
+		}
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"He4rt\\Message\\Tests\\": "tests/"
+		}
+	},
+	"minimum-stability": "stable",
+	"extra": {
+		"laravel": {
+			"providers": [
+				"He4rt\\Message\\Providers\\MessageServiceProvider"
+			]
+		}
+	}
+}
