+++ app-modules/message/src/Actions/PersistMessage.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace He4rt\Message\Actions;
+
+use He4rt\Message\DTOs\NewMessageDTO;
+use He4rt\Message\Entities\MessageEntity;
+use He4rt\Message\Repositories\MessageRepository;
+
+class PersistMessage
+{
+    public function __construct(
+        private readonly MessageRepository $messageRepository
+    ) {
+    }
+
+    public function handle(
+        NewMessageDTO $messageDTO,
+        int $obtainedExperience,
+        string $providerEntity
+    ): MessageEntity {
+        return $this->messageRepository->create($messageDTO, $providerEntity, $obtainedExperience);
+    }
+}
