+++ app-modules/message/src/Repositories/MessageRepository.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace He4rt\Message\Repositories;
+
+use He4rt\Message\DTOs\NewMessageDTO;
+use He4rt\Message\Entities\MessageEntity;
+
+interface MessageRepository
+{
+    public function create(NewMessageDTO $payload, string $providerId, int $obtainedExperience): MessageEntity;
+}
