+++ app-modules/message/src/Repositories/VoiceRepository.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace He4rt\Message\Repositories;
+
+use He4rt\Message\DTOs\NewVoiceMessageDTO;
+use He4rt\Message\Entities\VoiceEntity;
+
+interface VoiceRepository
+{
+    public function create(NewVoiceMessageDTO $messageDTO, string $providerId, int $obtainedExperience): VoiceEntity;
+}
